requirements.txt
streamlit==1.42.1
google-generativeai==0.8.4
gTTS==2.5.4
pandas==2.2.3
requests==2.32.3
Pillow==11.1.0
python-dotenv==1.0.1
